MAJOR: set a channel per engine-id
Each HAProxy process owns an engine ID. When you reload HAProxy, it will
compute a new engine ID.
The ASYNC mode of SPOE applies to all connections pointing to a same
engine-id.
Since the current implementation does not take into account the
engine-id when sending an ACK, it may be possible that an ACK is sent to
the wrong HAProxy client.

This patch fixes this issue by creating one channel per engine id, so
there is no chance to send an ACK to a wrong connection.