BUG/MAJOR: version-compare: fix comparison for minimum required and current program version
Fixed regex used to extract program version.
`version` arg was unused. Now, it is renamed as `minVer` and used in comparison.
Additional `compareVersion` func added to compare minVer with currVer.
Comparision with `strings.Compare` didn't worked for cases like `0.1.10 > 0.1.2`
and therefore it is removed.
Added unit test for `compareVersion` func.